requirements.txt
albumentations==1.4.10
filelock==3.15.3
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
pandas==2.2.2
pillow==10.3.0
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
smmap==5.0.1
streamlit==1.36.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.4.1
threadpoolctl==3.5.0
tifffile==2024.6.18
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torch-vision==0.1.6.dev0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1